requirements.txt
datasets
pandas
sentence-transformers
faiss-cpu
transformers
scikit-learn
streamlit
